requirements.txt
Django==1.9.5
djangorestframework==3.3.3
django-filter==0.13.0
gunicorn==19.4.5
psycopg2==2.6.1
dj-database-url==0.4.0
python-decouple==3.0
